getMesure du turfu OK

+++ En-tete/Capteur.h
@@ -10,15 +10,15 @@ class Capteur
 public:
   // Constructeurs
   Capteur();
-  Capteur(string sensorId, int longitude, int lattitude, string description);
+  Capteur(string sensorId, float longitude, float lattitude, string description);
 
   //Accesseurs et mutateurs
   void setSensorId(string sensorId);
-  void setLongitude(int longitude);
+  void setLongitude(float longitude);
   void setDescription(string description);
-  void setLattitude(int lattitude);
-  int getLongitude() const;
-  int getLattitude() const;
+  void setLattitude(float lattitude);
+  float getLongitude() const;
+  float getLattitude() const;
   string getSensorId() const;
   string getDescription() const;
 
@@ -29,8 +29,8 @@ public:
 
 private: 
   string sensorId;
-  int longitude;
-  int lattitude;
+  float longitude;
+  float lattitude;
   string description;
 };
